+++ bitmovinanalyticsexample/src/main/java/com/bitmovin/bitmovinanalyticsexample/MainActivity.java
@@ -18,6 +18,11 @@ import com.bitmovin.player.config.advertising.AdItem;
 import com.bitmovin.player.config.advertising.AdSource;
 import com.bitmovin.player.config.advertising.AdSourceType;
 import com.bitmovin.player.config.advertising.AdvertisingConfiguration;
+import com.bitmovin.player.config.track.AudioTrack;
+
+import com.bitmovin.player.config.track.SubtitleTrack;
+import java.util.Arrays;
+import java.util.List;
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener {
 
@@ -27,6 +32,8 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     private Button releaseButton;
     private Button createButton;
     private Button changeSource;
+    private Button changeAudio;
+    private Button changeSubtitle;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -38,6 +45,10 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         createButton.setOnClickListener(this);
         changeSource = findViewById(R.id.change_source);
         changeSource.setOnClickListener(this);
+        changeAudio = findViewById(R.id.change_audio);
+        changeAudio.setOnClickListener(this);
+        changeSubtitle = findViewById(R.id.change_subtitle);
+        changeAudio.setOnClickListener(this);
 
         this.bitmovinPlayerView = this.findViewById(R.id.bitmovinPlayerView);
 
@@ -68,8 +79,11 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     }
 
     protected BitmovinPlayerCollector initializeAnalytics() {
+        // This is an internal Bitmovin Analytics Key
+        private final String LOCAL_DEVELOPMENT_KEY = "17e6ea02-cb5a-407f-9d6b-9400358fbcc0";
+
         //Step 1: Create your analytics config object
-        BitmovinAnalyticsConfig bitmovinAnalyticsConfig = new BitmovinAnalyticsConfig("17e6ea02-cb5a-407f-9d6b-9400358fbcc0");
+        BitmovinAnalyticsConfig bitmovinAnalyticsConfig = new BitmovinAnalyticsConfig(LOCAL_DEVELOPMENT_KEY);
 
         //Step 2: Add optional parameters
         bitmovinAnalyticsConfig.setVideoId("androidVideoDASHStatic");
@@ -96,7 +110,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         SourceConfiguration sourceConfiguration = new SourceConfiguration();
 
         // Add a new source item
-        sourceConfiguration.addSourceItem("https://bitmovin-a.akamaihd.net/content/MI201109210084_1/mpds/f08e80da-bf1d-4e3d-8899-f0f6155f6efa.mpd");
+        sourceConfiguration.addSourceItem("https://bitmovin-a.akamaihd.net/content/sintel/hls/playlist.m3u8");
 
         return sourceConfiguration;
     }
@@ -134,6 +148,10 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
             initializeBitmovinPlayer();
         } else if (v == changeSource) {
             onPlayerChangeSource();
+        } else if (v == changeAudio) {
+            onAudioTrackChange();
+        } else if (v == changeSubtitle) {
+            onSubtitleChange();
         }
     }
 
@@ -141,6 +159,25 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         SourceConfiguration config = new SourceConfiguration();
         config.addSourceItem("http://bitdash-a.akamaihd.net/content/sintel/sintel.mpd");
         bitmovinPlayer.load(config);
+
+    }
+
+    private void onAudioTrackChange() {
+        AudioTrack[] available = bitmovinPlayer.getAvailableAudio();
+        List<AudioTrack> audioTracks = Arrays.asList(available);
+        int index = audioTracks.indexOf(bitmovinPlayer.getAudio());
+
+        String id = available[(index + 1) % available.length].getId();
+        bitmovinPlayer.setAudio(id);
+    }
+
+    private void onSubtitleChange() {
+        SubtitleTrack[] available = bitmovinPlayer.getAvailableSubtitles();
+        List<SubtitleTrack> audioTracks = Arrays.asList(available);
+        int index = audioTracks.indexOf(bitmovinPlayer.getSubtitle());
+
+        String id = available[(index + 1) % available.length].getId();
+        bitmovinPlayer.setSubtitle(id);
     }
 
     private static final String AD_SOURCE_1 = "https://pubads.g.doubleclick.net/gampad/ads?sz=640x480&iu=/124319096/external/single_ad_samples&ciu_szs=300x250&impl=s&gdfp_req=1&env=vp&output=vast&unviewed_position_start=1&cust_params=deployment%3Ddevsite%26sample_ct%3Dredirecterror&nofb=1&correlator=";