+++ collector/src/main/java/com/bitmovin/analytics/data/HttpBackend.java
@@ -23,13 +23,14 @@ public class HttpBackend implements Backend {
 
     @Override
     public void send(EventData eventData) {
-        Log.d(TAG, String.format("Sending sample: %s (state: %s, videoId: %s, startupTime: %d, videoStartupTime: %d, buffered: %d)",
+        Log.d(TAG, String.format("Sending sample: %s (state: %s, videoId: %s, startupTime: %d, videoStartupTime: %d, buffered: %d, audioLanguage: %s)",
                 eventData.getImpressionId(),
                 eventData.getVideoId(),
                 eventData.getState(),
                 eventData.getStartupTime(),
                 eventData.getVideoStartupTime(),
-                eventData.getBuffered()));
+                eventData.getBuffered(), 
+                eventData.getAudioLanguage()));
         this.httpClient.post(analyticsBackendUrl, DataSerializer.serialize(eventData), null);
     }