+++ collector/src/main/java/com/bitmovin/analytics/utils/ErrorUtil.kt
@@ -1,0 +1,6 @@
+package com.bitmovin.analytics.utils
+
+val Throwable.topOfStacktrace : Array<String>
+get() = this.stackTrace.take(10).map { it.toString() }.toTypedArray()
+
+